deps: Rewrite replaceModule to match by name _or_ path
Previously replaceModule only matched via name, which was appropriate for
updating `requires` and `children` but was not appropriate for `requiredBy`.
Because `requiredBy` is a list of modules that require this module it can
and will include modules that exist more than once in the physical tree but
the current node as a dependency.  As such, those modules need to be uniqued
on path instead.